# flo-device-service/src/docs/swagger.yaml
basePath: /v1
definitions:
...
  main.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  main.FirestoreToken:
    properties:
      token:
        example: tbd
        type: string
    type: object
...
host: flo-device-service.flocloud.co
info:
  contact:
    email: alex.galushka@flotechnologies.com
    name: Alexander Galushka
  description: This is service which takes provides information about FLO devices,
    its properties and states.
  license: {}
  title: Device Service API
  version: "1.0"
paths:
...
  /firestore/auth:
    post:
      consumes:
      - application/json
      description: |-
        generates custom JWT token to authenticate clients with Firestore limiting access to specified asset ids
        it is achieved by passing the ids of customer's asset such as devices in the request body which is going to be baked in as a custom claim "ids" into the jwt
      parameters:
      - description: specifies the devices ids client has access to
        in: body
        name: firestoreAuth
        required: true
        schema:
          $ref: '#/definitions/FirestoreAuth'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.FirestoreToken'
        "400":
          description: parameter ids has to be a string" "parameter ids can not be
            empty
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: failed to generate firestore jwt
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: generates custom JWT token to authenticate clients with Firestore
      tags:
      - devices
      - firestore
...
schemes:
- http
- https
swagger: "2.0"